Moved Pooling to embeddings/utils.py
