+++ llama_index/llms/huggingface.py
@@ -1,9 +1,6 @@
 import logging
-from enum import Enum
 from threading import Thread
 from typing import TYPE_CHECKING, Any, Callable, Dict, List, Optional, Sequence, Union
-
-import numpy as np
 
 from llama_index.bridge.pydantic import Field, PrivateAttr
 from llama_index.callbacks import CallbackManager
@@ -345,34 +342,6 @@ def conversational_output_to_chat_response(
     return ChatResponse(
         message=ChatMessage(role=role, content=output["generated_text"])
     )
-
-
-class Pooling(str, Enum):
-    """Enum of possible pooling choices with pooling behaviors."""
-
-    CLS = "cls"
-    MEAN = "mean"
-
-    def __call__(self, array: np.ndarray) -> np.ndarray:
-        if self == self.CLS:
-            return self.cls_pooling(array)
-        return self.mean_pooling(array)
-
-    @classmethod
-    def cls_pooling(cls, array: np.ndarray) -> np.ndarray:
-        if len(array.shape) == 3:
-            return array[:, 0]
-        if len(array.shape) == 2:
-            return array[0]
-        raise NotImplementedError(f"Unhandled shape {array.shape}.")
-
-    @classmethod
-    def mean_pooling(cls, array: np.ndarray) -> np.ndarray:
-        if len(array.shape) == 3:
-            return array.mean(axis=1)
-        if len(array.shape) == 2:
-            return array.mean(axis=0)
-        raise NotImplementedError(f"Unhandled shape {array.shape}.")
 
 
 class HuggingFaceInferenceAPI(LLM):
